requirements.txt
numpy==1.26.4
tensorflow==2.16.1  # Compatible with Python 3.12, not yet officially supported with Python 3.13
matplotlib==3.10.0
scikit-learn==1.6.1
flask==2.3.3
pillow==11.1.0
opencv-python==4.11.0.86
werkzeug==2.3.7
setuptools==69.0.2 